+++ rsocket-core/src/commonMain/kotlin/io/rsocket/kotlin/internal/Connect.kt
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package io.rsocket.kotlin.core
+package io.rsocket.kotlin.internal
 
 import io.rsocket.kotlin.*
-import io.rsocket.kotlin.internal.*
+import io.rsocket.kotlin.core.*
 
 @OptIn(TransportApi::class)
 internal suspend inline fun Connection.connect(
