+++ rsocket-core/src/jvmMain/kotlin/io/rsocket/kotlin/internal/currentMillis.kt
@@ -14,6 +14,6 @@
  * limitations under the License.
  */
 
-package io.rsocket.kotlin.keepalive
+package io.rsocket.kotlin.internal
 
 internal actual fun currentMillis(): Long = System.currentTimeMillis()
